+++ Runtime/Variables/Variable.cs
@@ -1,16 +1,19 @@
 ﻿using System;
+using Codetox.GameEvents;
 using UnityEngine;
 
 namespace Codetox.Variables
 {
     // TODO [#24]: Add documentation to Variable.cs
-    public abstract class Variable<T> : CustomScriptableObject
+    public abstract class Variable<T> : BaseVariable
     {
         [SerializeField] private T value;
 
         [SerializeField]
         [Tooltip("Prevent the value of this variable from being modified at runtime by other scripts.")]
         private bool readOnly;
+
+        public GameEvent<T> onValueChanged;
 
         public T Value
         {
@@ -20,20 +23,36 @@ namespace Codetox.Variables
                 if (readOnly) return;
                 if (this.value.Equals(value)) return;
                 this.value = value;
-                OnValueChanged?.Invoke(value);
+                if (onValueChanged) onValueChanged.Invoke(value);
             }
-        }
-
-        private void Reset()
-        {
-            Value = default;
         }
 
         private void OnEnable()
         {
-            Value = value;
+            if (onValueChanged) onValueChanged.Invoke(value);
         }
 
-        public event Action<T> OnValueChanged;
+        public override VariableData GetVariableData()
+        {
+            return new VariableData<T>(value);
+        }
+
+        public override void LoadVariableData(VariableData data)
+        {
+            var variableData = data as VariableData<T> ?? new VariableData<T>(default);
+            value = variableData.value;
+            if (onValueChanged) onValueChanged.Invoke(value);
+        }
+
+        [Serializable]
+        private class VariableData<TV> : VariableData
+        {
+            public TV value;
+
+            public VariableData(TV value)
+            {
+                this.value = value;
+            }
+        }
     }
 }