+++ Runtime/Variables/VariableData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Codetox.Variables
+{
+    [Serializable]
+    public abstract class VariableData
+    {
+    }
+}